Switch to floating point, camera relative position buffer

+++ gfx/gl/shaders/billboard.frag
@@ -10,7 +10,7 @@ uniform float size;
 
 include(`materialOut.glsl')
 
-flat in ivec3 ModelPos;
+flat in vec3 ModelPos;
 flat in float Yaw;
 flat in float Depth;
